Added swagger to dev properties. Double as amount changed on numeric(18,2) in sql, BigDecimal in code. SuccessfulReturn now only when data was found.

+++ src/main/java/com/example/employee/dto/JsonSalary.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 
 
 public record JsonSalary(@NotNull Long employeeId,
-                         @NotNull Double amount,
+                         @NotNull BigDecimal amount,
                          @NotNull LocalDate startDate) {
 
     public JsonSalary(Salary salary) {
